Implemented synchronised result in division module
It was required that the division occurs only when the user decides
this, i.e. after both the divided and divider have been setup. In other
words the quotient and remainder should not be recalculated each time
the user changes the divided or the divider.

For this purpose two additional attributes have been introduced:
- command: a write-only attribute that when setup triggers the
  calculation (synchronisation); can have two possible values: "divide"
(triggers calculation based on existing divided/divider values)
and "reset" (sets divided and divider to default values and updates
quotient and remainder too)
- status: a read-only attribute that contains the current status of the four numbers (divided,
  divider, quotient, remainder); can also have two possible values: "synced" (the quotient and remainder
reflect the divided and divider values; set after each successful
command execution) or "dirty" (gets setup when the divided and/or divider
change without any subsequent command execution; a successful command execution is required for setting it back to
"synced")

Note: an invalid or unsuccessful command leaves the status unchanged.

The application GreatestCommonDivisor also had to be updated
accordingly.